refactor(cockroach): share retry helpers
