+++ sqlspec/adapters/cockroach/__init__.py
@@ -1,0 +1,1 @@
+"""Shared CockroachDB adapter helpers."""
