+++ sqlspec/adapters/cockroach_psycopg/config.py
@@ -6,17 +6,14 @@ from psycopg import crdb as psycopg_crdb
 from psycopg_pool import AsyncConnectionPool, ConnectionPool
 from typing_extensions import NotRequired
 
+from sqlspec.adapters.cockroach._shared_core import CockroachRetryConfig
 from sqlspec.adapters.cockroach_psycopg._typing import (
     CockroachAsyncConnection,
     CockroachPsycopgAsyncSessionContext,
     CockroachPsycopgSyncSessionContext,
     CockroachSyncConnection,
 )
-from sqlspec.adapters.cockroach_psycopg.core import (
-    CockroachPsycopgRetryConfig,
-    apply_driver_features,
-    build_statement_config,
-)
+from sqlspec.adapters.cockroach_psycopg.core import apply_driver_features, build_statement_config
 from sqlspec.adapters.cockroach_psycopg.driver import (
     CockroachPsycopgAsyncDriver,
     CockroachPsycopgAsyncExceptionHandler,
@@ -197,7 +194,7 @@ class CockroachPsycopgSyncConfig(
         statement_config, driver_features = apply_driver_features(statement_config, driver_features)
 
         driver_features.setdefault("enable_auto_retry", True)
-        _ = CockroachPsycopgRetryConfig.from_features(driver_features)
+        _ = CockroachRetryConfig.from_features(driver_features)
 
         # Extract user connection hook before storing driver_features
         features_dict = dict(driver_features) if driver_features else {}
@@ -398,7 +395,7 @@ class CockroachPsycopgAsyncConfig(
         statement_config, driver_features = apply_driver_features(statement_config, driver_features)
 
         driver_features.setdefault("enable_auto_retry", True)
-        _ = CockroachPsycopgRetryConfig.from_features(driver_features)
+        _ = CockroachRetryConfig.from_features(driver_features)
 
         # Extract user connection hook before storing driver_features
         features_dict = dict(driver_features) if driver_features else {}
