+++ tests/unit/adapters/test_cockroach/__init__.py
@@ -1,0 +1,1 @@
+"""Shared CockroachDB adapter tests."""
